Split utils.dart into two files in dir utils/
