+++ lib/src/solid/utils/misc.dart
@@ -1,4 +1,4 @@
-/// Common utility functions used across the package.
+/// Miscellaneous utility functions used across the package.
 ///
 // Time-stamp: <Friday 2024-02-16 10:59:10 +1100 Graham Williams>
 ///
@@ -32,20 +32,18 @@ library;
 
 import 'dart:convert';
 
-import 'package:flutter/foundation.dart' hide Key;
-
 import 'package:crypto/crypto.dart';
 import 'package:encrypt/encrypt.dart';
-import 'package:fast_rsa/fast_rsa.dart';
 import 'package:jwt_decoder/jwt_decoder.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:path/path.dart' as path;
 import 'package:rdflib/rdflib.dart';
 import 'package:solid_auth/solid_auth.dart';
-import 'package:solid_auth/src/openid/openid_client.dart';
 
 import 'package:solidpod/src/solid/api/rest_api.dart';
 import 'package:solidpod/src/solid/constants.dart';
+import 'package:solidpod/src/solid/utils/authdata_manager.dart'
+    show AuthDataManager;
 
 // solid-encrypt uses unencrypted local storage and refers to http://yarrabah.net/ for predicates definition,
 // do not use it before it is updated (same as what the gurriny project does)
@@ -375,200 +373,6 @@ Future<void> saveWebId(String webId) async =>
 
 Future<String?> getWebId() async =>
     secureStorage.read(key: webIdSecureStorageKey);
-
-/// [AuthDataManager] is a class to manage auth data returned by
-/// solid-auth authenticate, including:
-/// - save auth data to secure storage
-/// - load auth data from secure storage
-/// - delete saved auth data from secure storage
-/// - refresh access token if necessary
-
-class AuthDataManager {
-  /// The URL for logging out
-  static String? _logoutUrl;
-
-  /// The RSA keypair and their JWK format
-  /// It seems Map<String, dynamic> does not work
-  static Map<dynamic, dynamic>? _rsaInfo;
-
-  /// The authentication response
-  static Credential? _authResponse;
-
-  /// The string key for storing auth data in secure storage
-  static const String _authDataSecureStorageKey = '_solid_auth_data';
-
-  /// Save the auth data returned by solid-auth authenticate in secure storage
-  /// It seems Map<String, dynamic> does not work
-  static Future<void> saveAuthData(Map<dynamic, dynamic> authData) async {
-    const keys = [
-      'client',
-      'rsaInfo',
-      'authResponse',
-      'tokenResponse',
-      'accessToken',
-      'idToken',
-      'refreshToken',
-      'expiresIn',
-      'logoutUrl',
-    ];
-
-    for (final key in keys) {
-      assert(authData.containsKey(key));
-    }
-
-    _logoutUrl = authData['logoutUrl'] as String;
-    _rsaInfo = authData['rsaInfo'] as Map<dynamic,
-        dynamic>; // Note that use Map<String, dynamic> does not seem to work
-    _authResponse = authData['authResponse'] as Credential;
-
-    await writeToSecureStorage(
-        _authDataSecureStorageKey,
-        jsonEncode({
-          'logout_url': _logoutUrl,
-          'rsa_info': jsonEncode({
-            ..._rsaInfo!,
-            // Overwrite the 'rsa' keypair in rsaInfo
-            'rsa': {
-              'public_key': _rsaInfo!['rsa'].publicKey as String,
-              'private_key': _rsaInfo!['rsa'].privateKey as String,
-            },
-          }),
-          'auth_response': _authResponse!.toJson(),
-        }));
-
-    debugPrint('AuthDataManager => saveAuthData() done');
-  }
-
-  /// Retrieve (and reconstruct) auth data from secure storage
-  /// It seems Map<String, dynamic> does not work
-  static Future<Map<dynamic, dynamic>?> loadAuthData() async {
-    if (_logoutUrl == null || _rsaInfo == null || _authResponse == null) {
-      final loaded = await _loadData();
-      if (!loaded) {
-        debugPrint('AuthDataManager => loadAuthData() failed');
-        return null;
-      }
-    }
-
-    assert(_logoutUrl != null && _rsaInfo != null && _authResponse != null);
-    try {
-      final tokenResponse = await _getTokenResponse();
-      return {
-        'client': _authResponse!.client,
-        'rsaInfo': _rsaInfo,
-        'authResponse': _authResponse,
-        'tokenResponse': tokenResponse,
-        'accessToken': tokenResponse!.accessToken,
-        'idToken': _authResponse!.idToken,
-        'refreshToken': _authResponse!.refreshToken,
-        'expiresIn': tokenResponse.expiresIn,
-        'logoutUrl': _logoutUrl,
-      };
-    } on Exception catch (e) {
-      debugPrint('AuthDataManager => loadAuthData() failed: $e');
-    }
-    return null;
-  }
-
-  /// Remove/delete auth data from secure storage
-  static Future<bool> removeAuthData() async {
-    try {
-      if (await secureStorage.containsKey(key: _authDataSecureStorageKey)) {
-        await secureStorage.delete(key: _authDataSecureStorageKey);
-        _logoutUrl = null;
-        _rsaInfo = null;
-        _authResponse = null;
-      }
-
-      return true;
-    } on Exception {
-      debugPrint('AuthDataManager => removeAuthData() failed');
-    }
-    return false;
-  }
-
-  /// Returns the (refreshed) access token
-  static Future<String?> getAccessToken() async {
-    final tokenResponse = await _getTokenResponse();
-    if (tokenResponse != null) {
-      return tokenResponse.accessToken;
-    } else {
-      debugPrint('AuthDataManager => getAccessToken() failed');
-    }
-    return null;
-  }
-
-  /// Returns the (updated) token response
-  static Future<TokenResponse?> _getTokenResponse() async {
-    if (_authResponse == null) {
-      final loaded = await _loadData();
-      if (!loaded) {
-        debugPrint('AuthDataManager => _getTokenResponse() failed');
-        return null;
-      }
-    }
-    assert(_authResponse != null);
-
-    try {
-      var tokenResponse = TokenResponse.fromJson(_authResponse!.response!);
-      if (JwtDecoder.isExpired(tokenResponse.accessToken!)) {
-        debugPrint(
-            'AuthDataManager => _getTokenResponse() refreshing expired token');
-        assert(_rsaInfo != null);
-        final rsaKeyPair = _rsaInfo!['rsa'] as KeyPair;
-        final publicKeyJwk = _rsaInfo!['pubKeyJwk'];
-        String tokenEndpoint =
-            _authResponse!.client.issuer.metadata['token_endpoint'] as String;
-        String dPopToken =
-            genDpopToken(tokenEndpoint, rsaKeyPair, publicKeyJwk, 'POST');
-        tokenResponse = await _authResponse!
-            .getTokenResponse(forceRefresh: true, dPoPToken: dPopToken);
-      }
-      return tokenResponse;
-    } on Exception catch (e) {
-      debugPrint('AuthDataManager => _getTokenResponse() failed: $e');
-    }
-    return null;
-  }
-
-  /// Returns the logout URL
-  static Future<String?> getLogoutUrl() async {
-    if (_logoutUrl == null) {
-      final loaded = await _loadData();
-      if (!loaded) {
-        debugPrint('AuthDataManager => getLogoutUrl() failed');
-        return null;
-      }
-    }
-    assert(_logoutUrl != null);
-    return _logoutUrl;
-  }
-
-  /// Reconstruct the rsaInfo from JSON string
-  static Map<dynamic, dynamic> _getRsaInfo(String rsaJson) {
-    final rsaInfo_ = jsonDecode(rsaJson) as Map<String, dynamic>;
-    final publicKey = rsaInfo_['rsa']['public_key'] as String;
-    final privateKey = rsaInfo_['rsa']['private_key'] as String;
-
-    return {...rsaInfo_, 'rsa': KeyPair(publicKey, privateKey)};
-  }
-
-  /// Retrieve auth data from secure storage
-  static Future<bool> _loadData() async {
-    final dataStr = await secureStorage.read(key: _authDataSecureStorageKey);
-
-    if (dataStr != null) {
-      final dataMap = jsonDecode(dataStr) as Map<String, dynamic>;
-      _logoutUrl = dataMap['logout_url'] as String;
-      _rsaInfo = _getRsaInfo(dataMap['rsa_info'] as String);
-      _authResponse =
-          Credential.fromJson((dataMap['auth_response'] as Map).cast());
-
-      return true;
-    }
-    return false;
-  }
-}
 
 /// Generates a list of default folder paths for a given application.
 ///
